+++ src/devices/neorotary4.rs
@@ -1,9 +1,11 @@
 use super::SeesawDeviceInit;
 use crate::{
-    modules::{encoder::EncoderModule, neopixel::NeopixelModule, status::StatusModule, HardwareId},
+    modules::{encoder::EncoderModule, status::StatusModule, HardwareId},
     prelude::GpioModule,
     seesaw_device, Driver, SeesawError,
 };
+#[cfg(feature = "module_neopixel")]
+use crate::modules::neopixel::NeopixelModule;
 
 seesaw_device! {
     /// Anecdotally, I've had a lot of issues with the quad rotary encoder.
@@ -22,6 +24,7 @@ impl<D: Driver> GpioModule<D> for NeoRotary4<D> {}
 impl<D: Driver> EncoderModule<D, 4> for NeoRotary4<D> {
     const ENCODER_BTN_PINS: [u8; 4] = [12, 14, 17, 9];
 }
+#[cfg(feature = "module_neopixel")]
 impl<D: Driver> NeopixelModule<D> for NeoRotary4<D> {
     type Color = NeoRotary4Color;
 
@@ -31,12 +34,13 @@ impl<D: Driver> NeopixelModule<D> for NeoRotary4<D> {
 
 impl<D: Driver> SeesawDeviceInit<D> for NeoRotary4<D> {
     fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
-        self.reset_and_verify_seesaw()
-            .and_then(|_| self.enable_neopixel())
-            .and_then(|_| self.enable_button(0))
-            .and_then(|_| self.enable_button(1))
-            .and_then(|_| self.enable_button(2))
-            .and_then(|_| self.enable_button(3))
-            .map(|_| self)
+        self.reset_and_verify_seesaw()?;
+        self.enable_button(0)?;
+        self.enable_button(1)?;
+        self.enable_button(2)?;
+        self.enable_button(3)?;
+        #[cfg(feature = "module_neopixel")]
+        self.enable_neopixel()?;
+        Ok(self)
     }
 }