+++ src/devices/neotrellis.rs
@@ -1,8 +1,10 @@
 use super::SeesawDeviceInit;
 use crate::{
-    modules::{keypad::KeypadModule, neopixel::NeopixelModule, status::StatusModule, HardwareId},
+    modules::{keypad::KeypadModule, status::StatusModule, HardwareId},
     seesaw_device, Driver, SeesawError,
 };
+#[cfg(feature = "module_neopixel")]
+use crate::modules::neopixel::NeopixelModule;
 
 seesaw_device! {
     name: NeoTrellis,
@@ -18,6 +20,7 @@ impl<D: Driver> KeypadModule<D> for NeoTrellis<D> {
     const NUM_ROWS: u8 = 4;
 }
 
+#[cfg(feature = "module_neopixel")]
 impl<D: Driver> NeopixelModule<D> for NeoTrellis<D> {
     type Color = NeoTrellisColor;
 
@@ -27,9 +30,10 @@ impl<D: Driver> NeopixelModule<D> for NeoTrellis<D> {
 
 impl<D: Driver> SeesawDeviceInit<D> for NeoTrellis<D> {
     fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
-        self.reset_and_verify_seesaw()
-            .and_then(|_| self.enable_neopixel())
-            .map(|_| self)
+        self.reset_and_verify_seesaw()?;
+        #[cfg(feature = "module_neopixel")]
+        self.enable_neopixel()?;
+        Ok(self)
     }
 }