+++ src/devices/rotary_encoder.rs
@@ -1,11 +1,13 @@
 use super::SeesawDeviceInit;
 use crate::{
     modules::{
-        encoder::EncoderModule, gpio::GpioModule, neopixel::NeopixelModule, status::StatusModule,
+        encoder::EncoderModule, gpio::GpioModule, status::StatusModule,
         HardwareId,
     },
     seesaw_device, Driver, SeesawError,
 };
+#[cfg(feature = "module_neopixel")]
+use crate::modules::neopixel::NeopixelModule;
 
 seesaw_device! {
   name: RotaryEncoder,
@@ -20,6 +22,8 @@ impl<D: Driver> GpioModule<D> for RotaryEncoder<D> {}
 impl<D: Driver> EncoderModule<D, 1> for RotaryEncoder<D> {
     const ENCODER_BTN_PINS: [u8; 1] = [24];
 }
+
+#[cfg(feature = "module_neopixel")]
 impl<D: Driver> NeopixelModule<D> for RotaryEncoder<D> {
     type Color = RotaryEncoderColor;
 
@@ -29,9 +33,10 @@ impl<D: Driver> NeopixelModule<D> for RotaryEncoder<D> {
 
 impl<D: Driver> SeesawDeviceInit<D> for RotaryEncoder<D> {
     fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
-        self.reset_and_verify_seesaw()
-            .and_then(|_| self.enable_button(0))
-            .and_then(|_| self.enable_neopixel())
-            .map(|_| self)
+        self.reset_and_verify_seesaw()?;
+        self.enable_button(0)?;
+        #[cfg(feature = "module_neopixel")]
+        self.enable_neopixel()?;
+        Ok(self)
     }
 }