+++ examples/generic_device_test.rs
@@ -1,7 +1,6 @@
 #![no_std]
 #![no_main]
 #![allow(incomplete_features)]
-#![feature(generic_const_exprs)]
 use adafruit_seesaw::{devices::GenericDevice, prelude::*, SeesawDriver};
 use cortex_m_rt::entry;
 use rtt_target::{rprintln, rtt_init_print};