+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project (hopefully) adheres to [Semantic Versioning](https://semver.org
 
 ## [Unreleased]
 
+
+## [0.11.1] - 2025-07-09
+
 ### Added
 
 - [#19](https://github.com/alexeden/adafruit-seesaw/pull/19) Add one feature per device and module, allowing disabling all and only enabling those needed. This allows building some of them on stable rust.